+++ demis/applications/chart-igs_service.tf
@@ -25,18 +25,20 @@ module "igs_service" {
 
   # Pass the values for the chart
   application_values = templatefile(local.igs_template_app, {
-    image_pull_secrets = var.pull_secrets,
-    repository         = var.docker_registry,
-    namespace          = var.target_namespace,
-    debug_enable       = var.debug_enabled,
-    istio_enable       = var.istio_enabled,
-    core_hostname      = var.core_hostname,
-    storage_hostname   = var.storage_hostname,
-    s3_storage_url     = local.s3_storage_url,
-    feature_flags      = try(var.feature_flags[local.igs_name], {}),
-    config_options     = try(var.config_options[local.igs_name], {}),
-    replica_count      = local.igs_replicas,
-    resource_block     = local.igs_resource_block
+    image_pull_secrets                  = var.pull_secrets,
+    repository                          = var.docker_registry,
+    namespace                           = var.target_namespace,
+    debug_enable                        = var.debug_enabled,
+    istio_enable                        = var.istio_enabled,
+    core_hostname                       = var.core_hostname,
+    storage_hostname                    = var.storage_hostname,
+    s3_storage_url                      = local.s3_storage_url,
+    feature_flags                       = try(var.feature_flags[local.igs_name], {}),
+    config_options                      = try(var.config_options[local.igs_name], {}),
+    replica_count                       = local.igs_replicas,
+    resource_block                      = local.igs_resource_block,
+    igs_encryption_certificate_checksum = try(kubernetes_secret.igs_encryption_certificate.metadata[0].annotations["checksum"], ""),
+    minio_secret_checksum               = try(kubernetes_secret.minio_credentials.metadata[0].annotations["checksum"], "")
   })
   istio_values = templatefile(local.igs_template_istio, {
     namespace                      = var.target_namespace,