+++ demis/applications/main.tf
@@ -20,6 +20,7 @@ locals {
     helm_repository_password      = var.helm_repository_password
     istio_routing_chart_version   = try(var.deployment_information["istio-routing"].main.version, "")
     istio_routing_chart_name      = coalesce(try(var.deployment_information["istio-routing"].chart-name, ""), "istio-routing")
+    reset_values                  = var.reset_values
   }
   # The version of the FHIR Profile Snapshots to use
   fhir_profile_snapshots = var.deployment_information["fhir-profile-snapshots"].main.version