+++ demis/applications/outputs.tf
@@ -18,6 +18,16 @@ output "dlp_enabled" {
   value       = try(local.dls_purger_information.enabled, false)
 }
 
+output "spp_ars_version" {
+  description = "The version of Surveillance-Pseudonym-Purger-ARS"
+  value       = local.spp_ars_enabled ? module.surveillance_pseudonym_purger_ars[0].app_chart_versions[0] : ""
+}
+
+output "spp_ars_enabled" {
+  description = "Whether Surveillance-Pseudonym-Purger-ARS is enabled"
+  value       = local.spp_ars_enabled
+}
+
 output "fhir_profile_snapshots" {
   description = "Version of the FHIR Profile Snapshots being used"
   value       = length(tolist(module.validation_service_core_metadata.current_profile_versions)) > 0 ? "[${join(", ", tolist(module.validation_service_core_metadata.current_profile_versions))}]" : local.fhir_profile_snapshots