+++ demis/applications/providers.tf
@@ -5,5 +5,10 @@ terraform {
       source  = "hashicorp/helm"
       version = ">= 3.0.0, < 4.0.0"
     }
+
+    kubernetes = {
+      source  = "hashicorp/kubernetes"
+      version = "2.38.0"
+    }
   }
 }