+++ ekm-template/applications/main.tf
@@ -18,5 +18,6 @@ locals {
     helm_repository_password      = var.helm_repository_password
     istio_routing_chart_version   = try(var.deployment_information["istio-routing"].main.version, "")
     istio_routing_chart_name      = coalesce(try(var.deployment_information["istio-routing"].chart-name, ""), "istio-routing")
+    reset_values                  = var.reset_values
   }
 }