+++ idm/applications/chart-keycloak.tf
@@ -1,5 +1,6 @@
 locals {
-  keycloak_name = "keycloak"
+  keycloak_name           = "keycloak"
+  tsl_deliverer_mock_name = "tsl-deliverer-mock"
   # Verify whether the service is defined or the deployment is explicitly enabled
   keycloak_enabled = contains(local.service_names, local.keycloak_name) ? var.deployment_information[local.keycloak_name].enabled : false
   # Check if stage-override templates are provided, otherwise use the project-defined ones
@@ -9,6 +10,18 @@ locals {
   keycloak_resources_overrides = try(var.resource_definitions[local.keycloak_name], {})
   keycloak_replicas            = lookup(local.keycloak_resources_overrides, "replicas", null) != null ? var.resource_definitions[local.keycloak_name].replicas : null
   keycloak_resource_block      = lookup(local.keycloak_resources_overrides, "resource_block", null) != null ? var.resource_definitions[local.keycloak_name].resource_block : null
+
+  # Determine whether to enable the TSL Deliverer Mock based on existence of tsl_deliverer_mock in deployment_information
+  enable_tsl_deliverer_mock = try(contains(keys(var.deployment_information), local.tsl_deliverer_mock_name) ? var.deployment_information[local.tsl_deliverer_mock_name].enabled : false, false)
+
+  # Fixed version with error handling and proper attribute reference
+  keycloak_index = try(
+    index(
+      [for cred in var.database_credentials : cred.secret-name],
+      "${local.keycloak_name}-database-secret"
+    ),
+    -1 # Default index if not found
+  )
 }
 
 module "keycloak" {
@@ -21,23 +34,31 @@ module "keycloak" {
   application_name       = local.keycloak_name
   deployment_information = var.deployment_information[local.keycloak_name]
   helm_settings          = local.common_helm_release_settings
-  depends_on             = [module.pgbouncer[0]]
+  depends_on             = [module.pgbouncer[0], module.gematik_idp[0]]
 
   # Pass the values for the chart
   application_values = templatefile(local.keycloak_template_app, {
-    image_pull_secrets = var.pull_secrets,
-    repository         = var.docker_registry,
-    namespace          = var.target_namespace,
-    istio_enable       = var.istio_enabled,
-    local_stage        = var.is_local_mode,
-    issuer_hostname    = local.auth_hostname,
-    data_version       = local.stage_configuration_data_version,
-    data_name          = local.stage_configuration_data_name,
-    feature_flags      = try(var.feature_flags[local.keycloak_name], {}),
-    config_options     = try(var.config_options[local.keycloak_name], {}),
-    replica_count      = local.keycloak_replicas,
-    resource_block     = local.keycloak_resource_block,
-    enable_import      = var.keycloak_user_import_enabled
+    image_pull_secrets                    = var.pull_secrets,
+    repository                            = var.docker_registry,
+    namespace                             = var.target_namespace,
+    istio_enable                          = var.istio_enabled,
+    local_stage                           = var.is_local_mode,
+    issuer_hostname                       = local.auth_hostname,
+    data_version                          = local.stage_configuration_data_version,
+    data_name                             = local.stage_configuration_data_name,
+    tsl_deliverer_mock_name               = local.tsl_deliverer_mock_name,
+    tsl_deliverer_mock_version            = local.tsl_deliverer_mock_version
+    feature_flags                         = try(var.feature_flags[local.keycloak_name], {}),
+    config_options                        = try(var.config_options[local.keycloak_name], {}),
+    replica_count                         = local.keycloak_replicas,
+    resource_block                        = local.keycloak_resource_block,
+    enable_import                         = var.keycloak_user_import_enabled,
+    enable_tsl_deliverer_mock             = local.enable_tsl_deliverer_mock,
+    keycloak_admin_account_checksum       = try(kubernetes_secret.keycloak_admin_account.metadata[0].annotations["checksum"], ""),
+    keycloak_portal_secret_checksum       = try(kubernetes_secret.keycloak_portal_secret.metadata[0].annotations["checksum"], ""),
+    keycloak_truststore_file_checksum     = try(kubernetes_secret.keycloak_truststore_file.metadata[0].annotations["checksum"], ""),
+    keycloak_truststore_password_checksum = try(kubernetes_secret.keycloak_truststore_password.metadata[0].annotations["checksum"], ""),
+    db_secret_checksum                    = try(kubernetes_secret.database_credentials[local.keycloak_index].metadata[0].annotations["checksum"], "")
   })
   istio_values = templatefile(local.keycloak_template_istio, {
     namespace       = var.target_namespace,