+++ idm/applications/chart-pgbouncer.tf
@@ -25,14 +25,16 @@ module "pgbouncer" {
 
   # Pass the values for the chart
   application_values = templatefile(local.pgbouncer_template_app, {
-    image_pull_secrets  = var.pull_secrets,
-    repository          = var.docker_registry,
-    database_host       = var.database_target_host,
-    istio_enable        = var.istio_enabled,
-    replica_count       = local.pgbouncer_replicas,
-    resource_block      = local.pgbouncer_resource_block
-    keycloak_db_enabled = local.keycloak_enabled
-    bundid_db_enabled   = local.bundid_enabled
+    image_pull_secrets           = var.pull_secrets,
+    repository                   = var.docker_registry,
+    database_host                = var.database_target_host,
+    istio_enable                 = var.istio_enabled,
+    replica_count                = local.pgbouncer_replicas,
+    resource_block               = local.pgbouncer_resource_block
+    keycloak_db_enabled          = local.keycloak_enabled
+    bundid_db_enabled            = local.bundid_enabled
+    postgres_tls_secret_checksum = try(kubernetes_secret.postgresql_tls_certificates.metadata[0].annotations["checksum"], ""),
+    userlist_secret_checksum     = try(kubernetes_secret.pgbouncer_userlist.metadata[0].annotations["checksum"], "")
   })
   istio_values = templatefile(local.pgbouncer_template_istio, {
     namespace = var.target_namespace