+++ idm/applications/chart-postgres.tf
@@ -9,6 +9,15 @@ locals {
   postgres_resources_overrides = try(var.resource_definitions[local.postgres_name], {})
   postgres_replicas            = lookup(local.postgres_resources_overrides, "replicas", null) != null ? var.resource_definitions[local.postgres_name].replicas : null
   postgres_resource_block      = lookup(local.postgres_resources_overrides, "resource_block", null) != null ? var.resource_definitions[local.postgres_name].resource_block : null
+
+  postgres_index = try(
+    index(
+      [for cred in var.database_credentials : cred.secret-name],
+      "${local.postgres_name}-secret"
+    ),
+    -1 # Default index if not found
+  )
+
 }
 
 module "postgres" {
@@ -24,16 +33,18 @@ module "postgres" {
 
   # Pass the values for the chart
   application_values = templatefile(local.postgres_template_app, {
-    image_pull_secrets  = var.pull_secrets,
-    repository          = var.docker_registry,
-    istio_enable        = var.istio_enabled,
-    debug_enable        = var.debug_enabled,
-    feature_flags       = try(var.feature_flags[local.postgres_name], {}),
-    config_options      = try(var.config_options[local.postgres_name], {}),
-    replica_count       = local.postgres_replicas,
-    resource_block      = local.postgres_resource_block
-    keycloak_db_enabled = local.keycloak_enabled
-    bundid_db_enabled   = local.bundid_enabled
+    image_pull_secrets           = var.pull_secrets,
+    repository                   = var.docker_registry,
+    istio_enable                 = var.istio_enabled,
+    debug_enable                 = var.debug_enabled,
+    feature_flags                = try(var.feature_flags[local.postgres_name], {}),
+    config_options               = try(var.config_options[local.postgres_name], {}),
+    replica_count                = local.postgres_replicas,
+    resource_block               = local.postgres_resource_block
+    keycloak_db_enabled          = local.keycloak_enabled
+    bundid_db_enabled            = local.bundid_enabled
+    postgres_tls_secret_checksum = try(kubernetes_secret.postgresql_tls_certificates.metadata[0].annotations["checksum"], ""),
+    db_secret_checksum           = try(kubernetes_secret.database_credentials[local.postgres_index].metadata[0].annotations["checksum"], "")
   })
   istio_values = templatefile(local.postgres_template_istio, {
     namespace = var.target_namespace